refactor: moveRule패키지 이름 컨벤션에 따라 소문자로 변경
